change data type to float32

+++ intel/onnx/lambda_function.py
@@ -38,7 +38,7 @@ def make_dataset(multipart_data, workload, framework):
         print(img)
         img = Image.open(img)
         img = img.resize((224,224), Image.ANTIALIAS)
-        img = np.array(img)
+        img = np.array(img).astype("float32")
         data = img.reshape(batch_size, channel, image_size, image_size)
 
         return data
